+++ Example12/src/Main.cpp
@@ -121,7 +121,7 @@ int main() {
             {{-0.5f, +0.5f, -0.5f}, {+0.0f, +1.0f}},// 35
     };
 
-    const std::vector<GLint> elements{
+    const std::vector<GLuint> elements{
             0, 1, 2, 3, 4, 5,      // 0
             6, 7, 8, 9, 10, 11,    // 1
             12, 13, 14, 15, 16, 17,// 2
@@ -148,7 +148,7 @@ int main() {
     GLuint EBO = 0;
     glGenBuffers(1, &EBO);
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, EBO);
-    glBufferData(GL_ELEMENT_ARRAY_BUFFER, static_cast<GLsizeiptr>(elements.size() * sizeof(GLfloat)), elements.data(), GL_STATIC_DRAW);
+    glBufferData(GL_ELEMENT_ARRAY_BUFFER, static_cast<GLsizeiptr>(elements.size() * sizeof(GLuint)), elements.data(), GL_STATIC_DRAW);
 
     GLuint texture;
     glGenTextures(1, &texture);
